% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wx_ACIS.R
\name{ACIS_elems}
\alias{ACIS_elems}
\title{See meteorological elements available for retrieval from \code{\link{get_wx_ACIS}}.}
\usage{
ACIS_elems()
}
\description{
Default is to retrieve daily maximum, minimum, and average temperatures,
 precipitation and snowfall totals, and snow depth. See also the
 \code{Optional parameters} documentation for the ACIS web server
 (\url{http://www.rcc-acis.org/docs_webservices.html}).
}
