% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrsmisc.R
\docType{package}
\name{adsmisc}
\alias{adsmisc}
\alias{adsmisc-package}
\title{\code{adsmisc} package}
\description{
Miscellaneous functions for life in the natural resources science, including
 retrieving weather (from multiple potential sources), tides, sunrise/sunset
 and moon face illumination for a particular geographic location. Also includes
 a few random functions to read tables from Access databases, manipulate spatial
 data, numeric vectors, and strings, etc.
}
