% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cap.R
\name{Cap}
\alias{Cap}
\title{Capitalize first or all words in string}
\usage{
Cap(string, words = c("all", "first"))
}
\arguments{
\item{string}{character vector}

\item{words}{which words to capitalize in string: \code{all} or \code{first} only?}
}
\description{
Capitalize first or all words in string
}
\examples{
Cap(c("adam d. smith", "northern parula", "title Caps please"))
Cap("this is a sentence.", "first")
}
