% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_16wind.R
\name{get_16wind}
\alias{get_16wind}
\title{Retrieve the point of a 16-wind compass rose for 0-360 degree bearings}
\usage{
get_16wind(degrees)
}
\arguments{
\item{degrees}{numeric or integer vector of compass bearings, in degrees and
constrained between 0 and 360 degrees}
}
\description{
This function converts compass bearings into one of 16 22.5 degree wide
 abbreviated text directions - specifically the eight principal winds
 (N, NE, E, SE, S, SW, W, NW) and eight half-winds that bisect the angles
 between the principal winds (NNE, ENE, ESE, SSE, SSW, WSW, WNW, NNW). The
 eight principal winds and the eight half-winds together form the 16-wind
 compass rose, with each compass point at a 22 1⁄2' angle from its neighbors
}
\examples{
get_16wind(90)
get_16wind(270)
get_16wind(22.5)
get_16wind(c(0, 45, 90, 135))
}
