% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACIS_meta.R
\name{get_ACIS_meta}
\alias{get_ACIS_meta}
\title{Retrieve date range of available meteorological data for one or more stations from
 the Applied Climate Information System (ACIS) stations}
\usage{
get_ACIS_meta(stn)
}
\arguments{
\item{stn}{vectors of station identification codes for which to retrieve the
period of available ACIS meteorological data. Several identification options
are available, such as WBAN, COOP, FAA, WMO, and GHCN. The available
\code{\link{find_wx_stns}} function returns the FAA code.}
}
\description{
SHubbard, K. G., A. T. DeGaetano, and K. D. Robbins. 2004. A modern Applied Climate
}
\examples{
\dontrun{
# See what's available around Mattamuskeet NWR
# When *not* returning station info
matta_wx_stns <- find_wx_stns(-76.175118, 35.451832, n_stns = 5)
get_ACIS_meta(matta_wx_stns) # Only 2 of 5 have ACIS data

# When returning station info as well
matta_wx_stns <- find_wx_stns(-76.175118, 35.451832, n_stns = 5, id_only = FALSE)
get_ACIS_meta(matta_wx_stns$id)
}
}
