% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buoy_wx.R
\name{get_buoy_wx}
\alias{get_buoy_wx}
\title{Retrieve historical METAR weather from NOAA buoy}
\usage{
get_buoy_wx(buoy = NULL, years = as.POSIXlt(Sys.Date())$year + 1900,
  out_tz = "GMT")
}
\arguments{
\item{buoy}{character or integer scalar of buoy from which to retrieve weather.
Browse at \url{http://www.ndbc.noaa.gov/} or see \link[rnoaa:buoy]{buoy_stations}.}

\item{years}{integer vector of years for which data are required}

\item{out_tz}{time zone specification to be used for output. System-specific
(see \link[base]{timezones}. Default is "GMT" (UTC; Universal Time,
Coordinated).}
}
\description{
Retrieve historical METAR weather from NOAA buoy; see
 \url{http://www.ndbc.noaa.gov/}. This function is a thin wrapper
 around \link[rnoaa]{buoy} mainly for multiple year queries and
 output formatting
}
\examples{
\dontrun{
wx <- get_buoy_wx("DPIA1", years = 2017, out_tz = "America/Chicago")
# Current year, in GMT
wx <- get_buoy_wx(41063)
}
}
