% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coop.R
\name{get_coop}
\alias{get_coop}
\title{Wrapper of rnoaa::coops_search that allows retrieval of more than
 one API-restricted request window at a time.}
\usage{
get_coop(begin_date = NULL, end_date = NULL, station_name = NULL,
  product, datum = NULL, units = "metric", time_zone = "gmt",
  application = "rnoaa", start = NULL, end = NULL, verbose = TRUE)
}
\arguments{
\item{begin_date}{(numeric) Date in yyyymmdd format. One of
\code{begin_date} or \code{start} is required.}

\item{end_date}{(numeric) Date in yyyymmdd format. One of
\code{end_date} or \code{end} is required.}

\item{station_name}{(numeric) Required station name/number. See
\url{https://tidesandcurrents.noaa.gov/map/} to locate station
name and view what data \code{product} might be available.}

\item{product}{(character) Specify the data type. See Details for options.
Required}

\item{datum}{(character) See below for Details. Required for all water
level products.}

\item{units}{(character) One of 'metric' (default) or 'english' (imperial)}

\item{time_zone}{(character) Time zone, one of 'gmt' (default), 'lst',
or 'lst_ldt'. See details at \code{\link[rnoaa:coops]{coops_search}}.}

\item{application}{(character) If called within an external package,
set to the name of your organization. Optional.}

\item{start}{character or Date scalar of start date ("YYYY-MM-DD")}

\item{end}{character or Date scalar of end date ("YYYY-MM-DD")}

\item{verbose}{(logical) print messages while processing requests?}
}
\value{
A \code{data.frame} with data or \code{NULL} if no data is found
}
\description{
This function is particularly useful for retrieving tide/water level data.
 Other functions in this package (e.g., \code{\link{get_wx}} or
 \code{\link{get_wx_ACIS}} may be better suited for basic meteorological data.
 All the products and options available at \code{\link[rnoaa:coops]{coops_search}}
 are available here except the ability to pass additional options to Curl.
 We also add add optional \code{start} and \code{end} date formats to be
 more consistent with other functions in this package. However, calls to
 \code{\link[rnoaa:coops]{coops_search}} without Curl options should work
 seemlessly with a call to \code{get_coop}. See example. Note, however,
 we return only the \code{data.frame} data component returned by
 \code{\link[rnoaa:coops]{coops_search}}
}
\details{
Options for the \code{product} paramater. Not all available at every station.
 One of:
\itemize{
 \item water_level - Preliminary or verified water levels, depending on
 availability
 \item air_temperature - Air temperature as measured at the station
 \item water_temperature - Water temperature as measured at the station
 \item wind - Wind speed, direction, and gusts as measured at the station
 \item air_pressure - Barometric pressure as measured at the station
 \item air_gap - Air Gap (distance between a bridge and the water's surface)
 at the station
 \item conductivity - The water's conductivity as measured at the station
 \item visibility - Visibility from the station's visibility sensor. A
 measure of atmospheric clarity
 \item humidity - Relative humidity as measured at the station
 \item salinity - Salinity and specific gravity data for the station
 \item one_minute_water_level - One minute water level data for the station
 \item predictions - 6 minute predictions water level data for the station
 \item hourly_height - Verified hourly height water level data for
 the station
 \item high_low - Verified high/low water level data for the station
 \item daily_mean - Verified daily mean water level data for the station
 \item monthly_mean - Verified monthly mean water level data for the station
 \item datums - datums data for the stations
 \item currents - Currents data for currents stations
}
Options for the datum parameter. Not all available for every station.
 One of:
\itemize{
 \item MHHW - Mean higher high water
 \item MHW - Mean high water
 \item MTL - Mean tide level
 \item MSL - Mean sea level
 \item MLW - Mean low water
 \item MLLW - Mean lower low water
 \item NAVD - North American Vertical Datum
 \item STND - Station datum
}
}
\examples{
\dontrun{
# Two months of water level data at Vaca Key (8723970)
# Fails with coops_search because query range is too long
coops_search(begin_date = 20140927, end_date = 20141115,
             station_name = 8723970, datum = "stnd",
             product = "water_level")
# Works with get_coop with exact same argument order
get_coop(begin_date = 20140927, end_date = 20141115,
         station_name = 8723970, datum = "stnd",
         product = "water_level")
}
}
