% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moon.R
\name{get_moon}
\alias{get_moon}
\title{Calculate moon exposure for a given time period}
\usage{
get_moon(datetime, lon, lat, exp_hrs = 6)
}
\arguments{
\item{datetime}{\link[base]{POSIXct} vector of datetimes from which moon
exposure should be calculated}

\item{lon}{numeric scalar of position longitude (decimal degrees; WGS84)}

\item{lat}{numeric scalar of position latitude (decimal degrees; WGS84)}

\item{exp_hrs}{integer of number of hours post-\code{datetime} to calculate
exposure}
}
\description{
This function calculates the proportion of the moon's face that is illuminated,
and whether the moon is above horizon at a given location, for a user-specified
number of hours from a provided \link[base]{POSIXct} object. Specifically, it
sums the proportion illuminated at the midpoint of every hour, conditional on
the moon being above local horizon. Thus, the returned value ranges from zero
(new moon or moon not above horizon) to \code{exp_hrs} (full moon above horizon
for the entire time from \code{datetime} to \code{datetime} + \code{exp_hrs}).
}
\examples{
get_moon(as.POSIXct(Sys.time()), -83, 34, 12)
# For fun, plot daily moon exposure for next lunar cycle
next_cycle <- as.POSIXct(seq.Date(Sys.Date(), by = 1, length.out = 28))
plot(next_cycle, get_moon(next_cycle, -83, 34, 24))
}
