% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sun.R
\name{get_sun}
\alias{get_sun}
\title{Retrieve sunrise or sunset times by geographic location}
\usage{
get_sun(lon, lat, start = Sys.Date(), end = start,
  direction = c("sunrise", "sunset", "dawn", "dusk"), out_tz = "GMT")
}
\arguments{
\item{lon}{numeric scalar of position longitude (decimal degrees; WGS84)}

\item{lat}{numeric scalar of position latitude (decimal degrees; WGS84)}

\item{start}{character or Date scalar of start date ("YYYY-MM-DD")}

\item{end}{character or Date scalar of start date ("YYYY-MM-DD")}

\item{direction}{any combination of "dawn", "dusk", "sunrise", or "sunset",
indicating which ephemerides should be calculated. Here, dawn and dusk
represent civil sunrise and civil sunset (sun 6 degrees below horizon),
respectively.}

\item{out_tz}{time zone specification to be used for output. System-specific
(see \link[base]{timezones}. Default is "GMT" (UTC; Universal Time,
Coordinated).}
}
\description{
Very thin wrapper around \link[maptools:sun-methods]{sunriset} mainly for the
 output formatting
}
\examples{
get_sun(-83, 34)
get_sun(-83, 34, start = "2017-12-31", end = "2017-12-31",
        direction = c("sunrise", "sunset"), out_tz = "America/New_York")
}
