% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UTM_zone.R
\name{get_UTM_zone}
\alias{get_UTM_zone}
\title{Retrieve UTM zone for a given position on the globe}
\usage{
get_UTM_zone(lon)
}
\arguments{
\item{lon}{numeric scalar of position longitude (decimal degrees; WGS84)}
}
\description{
There are likely some places where this fails but it works for every
place I've tried.
}
\examples{
get_UTM_zone(-83)
}
