% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wx.R
\name{get_wx}
\alias{get_wx}
\title{Retrieve METAR weather from a US ASOS station}
\usage{
get_wx(stn = NULL, start = Sys.Date(), end = start, out_tz = "GMT")
}
\arguments{
\item{stn}{character scalar of US ASOS station from which to retrieve
weather; see \code{\link{find_wx_stns}} using option \code{asos_only} == \code{TRUE}.}

\item{start}{character or Date scalar OR vector (see Details) of start date ("YYYY-MM-DD")}

\item{end}{character or Date scalar of end date ("YYYY-MM-DD"); optional (defaults
to same day as start)}

\item{out_tz}{time zone specification to be used for output. System-specific
(see \link[base]{timezones}. Default is "GMT" (UTC; Universal Time,
Coordinated).}
}
\description{
Retrieve METAR weather from a US ASOS station
}
\details{
A vector of dates can also be passed to the \code{start} argument if
 desired, in which case the \code{end} argument is ignored.
}
\examples{
get_wx("ATL", "2017-12-25")
get_wx("AHN", c("2017-11-01", "2017-12-01"))
}
