% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wx_ACIS.R
\name{get_wx_ACIS}
\alias{get_wx_ACIS}
\title{Retrieve meteorological data from the Applied Climate Information System (ACIS)}
\usage{
get_wx_ACIS(stn = NULL, start = Sys.Date(), end = start,
  elems = c("maxt", "mint", "avgt", "pcpn", "snow", "snwd"))
}
\arguments{
\item{stn}{scalar of station identification code for which to retrieve
meteorological data. Several options are available, such as WBAN, COOP, FAA, WMO,
ICAO, and GHCN. \code{\link{find_wx_stns}} returns one of the FAA, WBAN, or COOP
code, depending on the station. See the \code{Required parameters} documentation
for the ACIS web server for details. (\url{http://www.rcc-acis.org/docs_webservices.html})}

\item{start}{character or Date scalar of start date ("YYYY-MM-DD")}

\item{end}{character or Date scalar of end date ("YYYY-MM-DD"); optional (defaults
to same day as start)}

\item{elems}{character vector of meteorological elements to retrieve.
Default is to retrieve daily maximum, minimum, and average temperatures,
precipitation and snowfall totals, and snow depth. See \code{\link{ACIS_elems}} or
the \code{Optional parameters} documentation for the ACIS web server for additional
details and options (\url{http://www.rcc-acis.org/docs_webservices.html}). Custom
cooling, heating, and growing degree days, water equivalent of snow depth, and pan
evaporation not currently supported.}
}
\description{
SHubbard, K. G., A. T. DeGaetano, and K. D. Robbins. 2004. A modern Applied Climate
}
\examples{
# Atlanta Hartsfield-Jackson airport
# Use default code returned by function; in this case the FAA code
(atl_def <- find_wx_stns(-84.430876, 33.640139, n_stns = 1))
atl <- get_wx_ACIS("ATL", start = "2017-04-01", end = "2017-04-30")
# Use WMO code
atl2 <- get_wx_ACIS("72219", start = "2017-04-01", end = "2017-04-30")
identical(atl, atl2)
}
