% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhum.R
\name{relhum}
\alias{relhum}
\title{Convert air and dew point temperatures (Celsius) to relative humidity}
\usage{
relhum(tempC, dewptC)
}
\arguments{
\item{tempC}{numeric vector of air temperatures (in Celsius)}

\item{dewptC}{numeric vector of dewpoint temperatures (in Celsius)}
}
\description{
Uses the August-Roche-Magnus formula (Lawrence 2005, Bull. Amer.
 Meteor. Soc. 225-233); Values of A and B from Alduchov and Eskridge
 (1996, J App Meteor 35:601-609)
}
\examples{
relhum(13, 12.3)
relhum(3, -0.4)
}
