% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vec.R
\name{scale_vec}
\alias{scale_vec}
\title{Transform a vector of numeric values to any custom numeric scale}
\usage{
scale_vec(vec, scale = c(0, 1))
}
\arguments{
\item{vec}{integer or numeric vector of values to be scaled}

\item{scale}{integer or numeric vector of length 2 indicating the
lower and upper limits of the target scale}
}
\description{
Transform a vector of numeric values to any custom numeric scale
}
\examples{
# Convert numeric vector to the range 0-1Use FAA code
v <- rnorm(10)
scale_vec(v)
v <- runif(10)
scale_vec(v, scale = c(0, 100))
}
