% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_colors.R
\name{view_colors}
\alias{view_colors}
\title{Visualize hex code colors}
\usage{
view_colors(colors, direction = c("horizontal", "vertical"))
}
\arguments{
\item{colors}{character vector of hex color codes}

\item{direction}{character scalar indicating desired plotting direction
(default is horizontal). See details.}
}
\description{
Helpful for visualizing one or more colors (hex code)
}
\details{
For horizontal displays, the colors proceed left to right. For vertical
  displays, the colors proceed bottom to top.
}
\examples{
view_colors(c("#440154FF", "#3B528BFF", "#21908CFF", "#5DC863FF", "#FDE725FF"))
view_colors(c("#440154FF", "#3B528BFF", "#21908CFF", "#5DC863FF", "#FDE725FF"),
            direction = "vertical")
}
