% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wx_stns.R
\name{find_wx_stns}
\alias{find_wx_stns}
\title{Find nearest ASOS weather station by geographic location (USA only)}
\usage{
find_wx_stns(lon, lat, n_stns = 5, id_only = TRUE, plot = FALSE)
}
\arguments{
\item{lon}{numeric scalar of position longitude (decimal degrees; WGS84)}

\item{lat}{numeric scalar of position latitude (decimal degrees; WGS84)}

\item{n_stns}{how many nearest stations to return?}

\item{id_only}{logical; return only station abbreviations or station info too?}

\item{plot}{logical; generate plot showing station positions relative to input
location (default \code{FALSE})}
}
\description{
The returned station abbreviations can then be passed to \code{\link{get_wx}}
}
\examples{
find_wx_stns(-83, 34)
find_wx_stns(-83, 34, id_only = FALSE)
}
